+++ scripts/util.js
@@ -42,120 +42,6 @@ export function isString(obj) {
 }
 
 /**
- * Fast rounding for positive numbers
- * @param {number} n
- * @returns {number}
- */
-export function roundFastPositive(n) { return (n + 0.5) << 0; }
-
-/**
- * Bresenham line algorithm to generate pixel coordinates for a line between two points.
- * All coordinates must be positive or zero.
- * @param {number} x0   First coordinate x value
- * @param {number} y0   First coordinate y value
- * @param {number} x1   Second coordinate x value
- * @param {number} y1   Second coordinate y value
- * @testing
-Draw = CONFIG.GeometryLib.Draw
-let [t0, t1] = canvas.tokens.controlled
-pixels = bresenhamLine(t0.center.x, t0.center.y, t1.center.x, t1.center.y)
-for ( let i = 0; i < pixels.length; i += 2 ) {
-  Draw.point({ x: pixels[i], y: pixels[i + 1]}, { radius: 1 });
-}
- */
-export function bresenhamLine(x0, y0, x1, y1) {
-  x0 = Math.round(x0);
-  y0 = Math.round(y0);
-  x1 = Math.round(x1);
-  y1 = Math.round(y1);
-
-  const dx = Math.abs(x1 - x0);
-  const dy = Math.abs(y1 - y0);
-  const sx = (x0 < x1) ? 1 : -1;
-  const sy = (y0 < y1) ? 1 : -1;
-  let err = dx - dy;
-
-  const pixels = [x0, y0];
-  while ( x0 !== x1 || y0 !== y1 ) {
-    const e2 = err * 2;
-    if ( e2 > -dy ) {
-      err -= dy;
-      x0 += sx;
-    }
-    if ( e2 < dx ) {
-      err += dx;
-      y0 += sy;
-    }
-
-    pixels.push(x0, y0);
-  }
-  return pixels;
-}
-
-export function* bresenhamLineIterator(x0, y0, x1, y1) {
-  x0 = Math.floor(x0);
-  y0 = Math.floor(y0);
-  x1 = Math.floor(x1);
-  y1 = Math.floor(y1);
-
-  const dx = Math.abs(x1 - x0);
-  const dy = Math.abs(y1 - y0);
-  const sx = (x0 < x1) ? 1 : -1;
-  const sy = (y0 < y1) ? 1 : -1;
-  let err = dx - dy;
-  yield { x: x0, y: y0 };
-  while ( x0 !== x1 || y0 !== y1 ) {
-    const e2 = err * 2;
-    if ( e2 > -dy ) {
-      err -= dy;
-      x0 += sx;
-    }
-    if ( e2 < dx ) {
-      err += dx;
-      y0 += sy;
-    }
-
-    yield { x: x0, y: y0 };
-  }
-}
-
-/**
- * Trim line segment to its intersection points with a rectangle.
- * If the endpoint is inside the rectangle, keep it.
- * Note: points on the right or bottom border of the rectangle do not count b/c we want the pixel positions.
- * @param {PIXI.Rectangle} rect
- * @param {Point} a
- * @param {Point} b
- * @returns { Point[2]|null } Null if both are outside.
- */
-export function trimLineSegmentToPixelRectangle(rect, a, b) {
-  rect = new PIXI.Rectangle(rect.x, rect.y, rect.width - 1, rect.height - 1);
-
-  if ( !rect.lineSegmentIntersects(a, b, { inside: true }) ) return null;
-
-  const ixs = rect.segmentIntersections(a, b);
-  if ( ixs.length === 2 ) return ixs;
-  if ( ixs.length === 0 ) return [a, b];
-
-  // If only 1 intersection:
-  //   1. a || b is inside and the other is outside.
-  //   2. a || b is on the edge and the other is outside.
-  //   3. a || b is on the edge and the other is inside.
-  // Point on edge will be considered inside by _getZone.
-
-  // 1 or 2 for a
-  const aOutside = rect._getZone(a) !== PIXI.Rectangle.CS_ZONES.INSIDE;
-  if ( aOutside ) return [ixs[0], b];
-
-  // 1 or 2 for b
-  const bOutside = rect._getZone(b) !== PIXI.Rectangle.CS_ZONES.INSIDE;
-  if ( bOutside ) return [a, ixs[0]];
-
-  // 3. One point on the edge; other inside. Doesn't matter which.
-  return [a, b];
-}
-
-/**
  * From https://stackoverflow.com/questions/14446511/most-efficient-method-to-groupby-on-an-array-of-objects
  * Takes an Array<V>, and a grouping function,
  * and returns a Map of the array grouped by the grouping function.
@@ -177,4 +63,49 @@ export function groupBy(list, keyGetter) {
     else collection.push(item);
   });
   return map;
+}
+
+
+/**
+ * Get the grid shape for a given set of grid coordinates.
+ * @param {number[2]} gridCoords      Array of [row, col] grid coordinates. See canvas.grid.grid.getGridPositionFromPixels
+ * @returns {PIXI.Rectangle|PIXI.Polygon}
+ */
+export function gridShapeFromGridCoords(gridCoords) {
+  const [tlx, tly] = canvas.grid.grid.getPixelsFromGridPosition(gridCoords[0], gridCoords[1]);
+  if ( canvas.grid.isHex ) return hexGridShape(tlx, tly);
+  return squareGridShape(tlx, tly)
+
+}
+
+/**
+ * Get a square grid shape from the top left corner position.
+ * @param {number} tlx      Top left x coordinate
+ * @param {number} tly      Top left y coordinate
+ * @returns {PIXI.Rectangle}
+ */
+function squareGridShape(tlx, tly) {
+  // Get the top left corner
+  const { w, h } = canvas.grid;
+  return new PIXI.Rectangle(tlx, tly, w, h);
+}
+
+/**
+ * Get a hex grid shape from the top left corner position.
+ * @param {number} tlx      Top left x coordinate
+ * @param {number} tly      Top left y coordinate
+ * @returns {PIXI.Polygon}
+ */
+function hexGridShape(tlx, tly, { width = 1, height = 1 } = {}) {
+  // Canvas.grid.grid.getBorderPolygon will return null if width !== height.
+  if ( width !== height ) return null;
+
+  // Get the top left corner
+  const points = canvas.grid.grid.getBorderPolygon(width, height, 0);
+  const pointsTranslated = [];
+  const ln = points.length;
+  for ( let i = 0; i < ln; i += 2) {
+    pointsTranslated.push(points[i] + tlx, points[i+1] + tly);
+  }
+  return new PIXI.Polygon(pointsTranslated);
 }